Trained a classifier and set up a detection. catsnout.xml is the haar classifier for catsnouts trained with opencv.
The classifier is used to look for catsnouts in each image taken.

If a catsnout is found:
The area of the catsnout and below is blurred and then thresholded to
get the biggest contour in this area (usually the catsnout and pray).
If this contour is reaching down below the catsnout rectangle and
takes up more pixels than a configurable threshold a pray is detected.

If the contour does not reach down, or the area taken up is below the
threshold, no pray is detected.

Or there can be no catsnout found at all.

The doorlock is operated on a trust value.
The thre possible detection outcomes each have a configurable trust factor assigned. e.g.
- catsnout detected, no pray       - factor 2.0
- catsnout detected, pray detected - factor 0.33
- no catsnout detected             - factor 0.95
The momentary trust value is multiplied with the detection outcome factor.
Starting value is 1.0
The door opens if the trust value is above or equal to 4.0